Refactor Features
In order to migrate to a single "Balthisar Tidy" with a one-time in-app purchase to
enable pro features, the current hard-coded features have to be refactored so that
features can be enabled in software rather than compiled in. This is tricky for
AppleScript and our helper application, because the dictionary will simply be present
making AppleScript available.

- All menus now use "Balthisar Tidy," and we will start to remove "for Work" branding.

- Preferences controller demoted to an object that has a MASPreferencesController, so
  that we can re-build the controller at will to add/remove panels.